Remove peerDependencies in favor of normal Dependencies.
This allows the user to just install the base package without having to manually install the peer dependencies. This also has the benefit of us having control over the entire package. If we decide to update some dependencies the user just needs to update our package and not update their own dependencies.